Reactify particles component
