+++ www/src/_data/particles.js
@@ -1,9 +1,11 @@
-const glob = require("glob")
-const path = require("path")
+const glob = require("glob");
+const path = require("path");
 
-const files = glob.sync(path.join(__dirname, "../images/particles/**/*.svg"))
-const images = files.map(file => path.basename(file, path.extname(file)))
+const files = glob.sync(
+  path.join(__dirname, "../assets/images/particles/**/*.svg")
+);
+const images = files.map((file) => path.basename(file, path.extname(file)));
 
 module.exports = {
-  images: process.env.DISABLE_PARTICLES ? [] : images
-}
+  images: process.env.DISABLE_PARTICLES ? [] : images,
+};
