--- www/src/_includes/components/particles/particles.template.njk
@@ -1,9 +1,0 @@
-<canvas id="particles-canvas" class="top-0 left-0 absolute" data-count="{{ count }}" data-images="{{ images }}"></canvas>
-
-<script>
-  onInit(function () {
-    App
-      .Particles
-      .init()
-  })
-</script>
