--- www/src/_includes/components/particles/particles.transformer.js
@@ -1,10 +1,0 @@
-const defaults = {
-  count: 10
-}
-
-module.exports = ({ count, ...props }) => {
-  return {
-    ...props,
-    count: count || defaults.count
-  }
-}
